+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## 0.14.3
 
+* `Common`:
+    * New type `Progress`
+* `Ktor`:
+    * `Client`:
+        * New universal `uniUpload` extension for `HttpClient`
+    * `Server`:
+        * New universal `handleUniUpload` extension for `ApplicationCall`
+
 ## 0.14.2
 
 * `Versions`: